added macro to check enums added some missing CGNS_ENUMV fixed some compiler warnings
git-svn-id: http://svn.code.sf.net/p/cgns/code/cgns/trunk@546 59ee3124-ab72-4556-8a47-5b88dfbcc004

+++ src/cgnstools/utilities/vgrid_to_cgns.c
@@ -544,7 +544,8 @@ int main (int argc, char *argv[])
     printf("reading volume grid from \"%s\"\n", basename);
     fflush(stdout);
     read_cogsg(basename);
-    printf("read %d coordinates and %d tetrahedra\n", nCoords, nTets);
+    printf("read %ld coordinates and %ld tetrahedra\n",
+        (long)nCoords, (long)nTets);
 
     /* read surface grid */
